requirements.txt
numpy==2.3.4
openai_whisper==20250625
python-dotenv==1.2.1
Requests==2.32.5
torch==2.9.0
torchaudio==2.9.0
